fix: use custom hook to handle calls to Reports API
